redo

+++ src/services/auth.js
@@ -1,16 +1,16 @@
 import createHttpError from 'http-errors';
 import { UsersCollection } from '../db/models/user.js';
 import bcrypt from 'bcrypt';
-import { SessionsCollection } from '../db/models/session.js';
 import { randomBytes } from 'crypto';
 import { FIFTEEN_MINUTES, ONE_DAY } from '../constants/index.js';
+import { SessionsCollection } from '../db/models/session.js';
 
 export const registerUser = async (payload) => {
-  const user = await UsersCollection.findOne({
-    email: payload.email,
-  });
+  const user = await UsersCollection.findOne({ email: payload.email });
   if (user) throw createHttpError(409, 'Email in use');
+
   const encryptedPassword = await bcrypt.hash(payload.password, 10);
+
   return await UsersCollection.create({
     ...payload,
     password: encryptedPassword,
@@ -26,8 +26,8 @@ export const loginUser = async (payload) => {
   if (!isEqual) {
     throw createHttpError(401, 'Unauthorized');
   }
-
   await SessionsCollection.deleteOne({ userId: user._id });
+
   const accessToken = randomBytes(30).toString('base64');
   const refreshToken = randomBytes(30).toString('base64');
 
@@ -56,15 +56,11 @@ const createSession = () => {
   };
 };
 
-export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
+export const refreshUserSession = async ({ sessionId, refreshToken }) => {
   const session = await SessionsCollection.findOne({
     _id: sessionId,
     refreshToken,
   });
-
-  if (!session) {
-    throw createHttpError(401, 'Session not found');
-  }
 
   const isSessionTokenExpired =
     new Date() > new Date(session.refreshTokenValidUntil);
@@ -73,9 +69,13 @@ export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
     throw createHttpError(401, 'Session token expired');
   }
 
-  const newSession = createSession();
+  if (!session) {
+    throw createHttpError(401, 'Session token expired or invalid');
+  }
 
   await SessionsCollection.deleteOne({ _id: sessionId, refreshToken });
+
+  const newSession = createSession();
 
   return await SessionsCollection.create({
     userId: session.userId,
